+++ plugin.video.kodipopcorntime/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.kodipopcorntime" name="KODI Popcorn Time" version="1.6.5" provider-name="Markop159">
+<addon id="plugin.video.kodipopcorntime" name="KODI Popcorn Time" version="1.6.6" provider-name="Markop159">
     <requires>
         <import addon="xbmc.python" version="2.14.0"/>
     </requires>
@@ -16,7 +16,7 @@
         <summary lang="en"><![CDATA[Immediately stream movies, TV Shows and Animes.]]></summary>
         <summary lang="es"><![CDATA[Immediately stream movies, TV Shows and Animes.]]></summary>
         <summary lang="fr"><![CDATA[Immediately stream movies, TV Shows and Animes.]]></summary>
-	<summary lang="hr"><![CDATA[Trenutačno gledajte filmove, TV Serije i Anime.]]></summary> 
+	      <summary lang="hr"><![CDATA[Trenutačno gledajte filmove, TV Serije i Anime.]]></summary>
         <summary lang="zh"><![CDATA[Immediately stream movies, TV Shows and Animes.]]></summary>
         <summary lang="nl"><![CDATA[Immediately stream movies, TV Shows and Animes.]]></summary>
 	      <summary lang="pl"><![CDATA[Immediately stream movies, TV Shows and Animes.]]></summary>
@@ -27,7 +27,7 @@
         <description lang="en"><![CDATA[With KODI Popcorn Time you can search for movies, TV shows and Animes that you can see immediately in KODI. NOTE: KODI Popcorn Time will continue sharing content while watching a movie TV show or Anime.]]></description>
         <description lang="es"><![CDATA[With KODI Popcorn Time you can search for movies, TV shows and Animes that you can see immediately in KODI. NOTE: KODI Popcorn Time will continue sharing content while watching a movie TV show or Anime.]]></description>
         <description lang="fr"><![CDATA[With KODI Popcorn Time you can search for movies, TV shows and Animes that you can see immediately in KODI. NOTE: KODI Popcorn Time will continue sharing content while watching a movie TV show or Anime.]]></description>
-	<description lang="hr"><![CDATA[S KODI Popcorn Time možete pretraživati filmove, TV serije i Anime koje možete odmah gledati u KODIJU. NAPOMENA: KODI Popcorn Time će nastaviti dijeliti sadržaj dok gledate film TV seriju ili Anime.]]></description>
+        <description lang="hr"><![CDATA[S KODI Popcorn Time možete pretraživati filmove, TV serije i Anime koje možete odmah gledati u KODIJU. NAPOMENA: KODI Popcorn Time će nastaviti dijeliti sadržaj dok gledate film TV seriju ili Anime.]]></description>
         <description lang="zh"><![CDATA[With KODI Popcorn Time you can search for movies, TV shows and Animes that you can see immediately in KODI. NOTE: KODI Popcorn Time will continue sharing content while watching a movie TV show or Anime.]]></description>
         <description lang="nl"><![CDATA[With KODI Popcorn Time you can search for movies, TV shows and Animes that you can see immediately in KODI. NOTE: KODI Popcorn Time will continue sharing content while watching a movie TV show or Anime.]]></description>
         <description lang="pl"><![CDATA[With KODI Popcorn Time you can search for movies, TV shows and Animes that you can see immediately in KODI. NOTE: KODI Popcorn Time will continue sharing content while watching a movie TV show or Anime.]]></description>
